Use data.get() rather than data['key'] in ratings.py
In one project, the call for setup() is conditionally called using
```
if __name__ == '__main__':
    setup(...)
```

These changes allow pyroma to finish without exception, and given the many "missing" errors and poor rating, to better discern the root cause of the problem -- that pyroma can not handle such case and was unable to discover the setup() call.

(Honestly I don't know this project did this! I think this is simply legacy carry-over, perhaps to allow some tool to import this file as a module without side-effect)