Do not complain that the version number should be a string, when it is a basestring.
This can happen when you use a Python 2.7 pyroma to check a package
that works on Python 2 and 3.  For example zest.releaser (5.0).  It
has 'from __future__ import unicode_literals' in its setup.py and has
version = '5.0'.  A Python 3 pyroma thinks this is fine, a Python 2.7
pyroma not: it is unicode where it expects a str.  But basestrings are
fine, and that is how the long_description is tested too.